requirements.txt
# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.8.0 

# Data processing
pandas>=2.1.4

# OpenAI API
openai>=1.41.0

# Environment variables
python-dotenv>=1.0.0

# Excel processing
openpyxl==3.1.2